pyproject.toml
[project]
name = "mcp-walkthrough"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.0",
    "fastmcp>=3.0.2",
    "nest-asyncio>=1.6.0",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
]
